Lint fixes, move listener
