+++ src/hooks/DAO/loaders/governance/useERC20LinearStrategy.ts
@@ -109,25 +109,5 @@ export const useERC20LinearStrategy = () => {
     };
   }, [ozLinearVotingContractAddress, action, baseContracts]);
 
-  useEffect(() => {
-    if (!azoriusContractAddress || !baseContracts) {
-      return;
-    }
-    const azoriusContract =
-      baseContracts.fractalAzoriusMasterCopyContract.asProvider.attach(azoriusContractAddress);
-
-    const timeLockPeriodFilter = azoriusContract.filters.TimelockPeriodUpdated();
-    const timelockPeriodListener: TypedListener<TimelockPeriodUpdatedEvent> = timelockPeriod => {
-      action.dispatch({
-        type: FractalGovernanceAction.UPDATE_TIMELOCK_PERIOD,
-        payload: BigNumber.from(timelockPeriod),
-      });
-    };
-    azoriusContract.on(timeLockPeriodFilter, timelockPeriodListener);
-    return () => {
-      azoriusContract.off(timeLockPeriodFilter, timelockPeriodListener);
-    };
-  }, [azoriusContractAddress, action, baseContracts]);
-
   return loadERC20Strategy;
 };
